add parser test for nonunique without cardinality, remove Apollo parser test and corresponding models
